+++ poc/CHANGES.md
@@ -1,4 +1,14 @@
 # Changes in the FPL Parser and Interpreter
+## 1.10.3
+* Refactoring:
+  * Adding the attribute 'reference' to the AuxSTVariable node in the symbol table
+  * Replacing AuxSTPredicateWithArgs outlined as 'var' to AuxSTVariable in the symbol table
+  * Replacing AuxSTPredicate outlined as 'undefined' to the new node type AuxSTUndefined in the symbol table
+  * Replacing AuxSTPredicate outlined as 'id' by a new node type AuxSTIdentifier in the symbol table
+  * Replacing AuxSTPredicate outlined as 'indexValue' and 'variadicVar' (nested into each other) by new node class AuxSTVariableVariadic and AuxSTInt
+  * Initial implementation of AuxSTUndefined, AuxSTIdentifier, AuxSTVariableVariadic, and AuxSTInt
+  * Renaming InbuiltIndex type to InbuiltVariableVariadic
+  * Adjusting the affected test cases of the symbol table
 ## 1.10.2
 * Bugfix:
   * Fixing infinite recursion in AuxSTSelf evaluation